reservoir correction

+++ mqtt_network/reservoir/reservoir_sensor.c
@@ -4,9 +4,9 @@
 /*   The following code is just a simulation of the output of a level sensor and the corresponding actuator that
      puts or fetches the water from the reservoir*/
 
-static double sensed_level=MAX_LEVEL;
+static int sensed_level=MAX_LEVEL;
 
-static double simulate_level(){
+static int simulate_level(){
     return sensed_level;
 }
